+++ packages/rake-db/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rake-db
 
+## 2.8.3
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.11.4
+  - orchid-core@0.4.1
+
 ## 2.8.2
 
 ### Patch Changes